LocalIntel now works Various fixes for KeepstarWatcher

+++ BotLibPlugins/GridScout/LocalPilot.cs
@@ -3,7 +3,7 @@ namespace GridScout2
   public record LocalPilot
   {
     public required string Name { get; init; }
-    public required long CharacterID { get; init; }
+    public required int CharacterID { get; init; }
     public required string StandingHint { get; init; }
   }
 }
